+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [3.2.1] - 2025-10-22
+
+### Changed
+- Improved release process to push package to nuget.org
+  ([cyberark/conjur-api-dotnet#17](https://github.com/cyberark/conjur-api-dotnet/issues/17), CNJR-9543)
+
 ## [3.2.0] - 2025-10-07
 
 ### Added